+++ lib/src/src.dart
@@ -1,3 +1,4 @@
+export 'caching/caching.dart';
 export 'persistence/persistence.dart';
 export 'state_management/state_management.dart';
 export 'widgets/widgets.dart';